+++ examples/widgets/pubspec.yaml
@@ -1,9 +1,12 @@
-name: widgets
+name: sky_widgets_examples
 dependencies:
   flutter: ">=0.0.3 <0.1.0"
   sky_tools: any
+  flutter_rendering_examples: any
 dependency_overrides:
   material_design_icons:
     path: ../../sky/packages/material_design_icons
   flutter:
     path: ../../sky/packages/sky
+  flutter_rendering_examples:
+    path: ../rendering
